# Remove one second waiting in hidding elements