--- a/module/VisualCeption.php
+++ b/module/VisualCeption.php
@@ -382,7 +382,9 @@ class VisualCeption extends CodeceptionModule
         foreach ($excludeElements as $element) {
             $this->hideElement($element);
         }
-        $this->webDriverModule->waitForElementNotVisible(array_pop($excludeElements));
+        if (!empty($excludeElements)) {
+            $this->webDriverModule->waitForElementNotVisible(array_pop($excludeElements));
+        }
     }
 
     /**
@@ -395,7 +397,9 @@ class VisualCeption extends CodeceptionModule
         foreach ($excludeElements as $element) {
             $this->showElement($element);
         }
-        $this->webDriverModule->waitForElementVisible(array_pop($excludeElements));
+        if (!empty($excludeElements)) {
+            $this->webDriverModule->waitForElementVisible(array_pop($excludeElements));
+        }
     }
 
     /**