And tests are faster :)

**before:**
Codeception PHP Testing Framework v2.2.9
Powered by PHPUnit 5.7.15 by Sebastian Bergmann and contributors.

Acceptance Tests (7) 
✔</SimpleCept: Check visual changes inside element (9.36s)
✔</NotSameSizeCest: See visual changes after size changes (8.09s)
✔</TimeComparisonCest: See visual changes (8.03s)
✔</TimeComparisonCest: Dont see visual changes (8.21s)
✔</TimeComparisonCest: See visual changes and hide element (8.20s)
✔</TimeComparisonCest: Dont see visual changes and hide element (8.22s)
✔</WriteCurrentImageCest: Write current image file (3.68s)

Time: 55.46 seconds, Memory: 13.75MB

OK (7 tests, 6 assertions)


**after:**
Codeception PHP Testing Framework v2.2.9
Powered by PHPUnit 5.7.15 by Sebastian Bergmann and contributors.

Acceptance Tests (7)
✔</SimpleCept: Check visual changes inside element (5.07s)
✔</NotSameSizeCest: See visual changes after size changes (4.06s)
✔</TimeComparisonCest: See visual changes (4.01s)
✔</TimeComparisonCest: Dont see visual changes (4.27s)
✔</TimeComparisonCest: See visual changes and hide element (4.57s)
✔</TimeComparisonCest: Dont see visual changes and hide element (4.58s)
✔</WriteCurrentImageCest: Write current image file (1.68s)

Time: 30.17 seconds, Memory: 13.75MB

OK (7 tests, 6 assertions)