@RayRom 2017-03-17
Not the right decision!!! In case of hiding of blocks, time for redraw of the page is necessary, that is possible do, reduce an interval to 1/4 seconds. You didn't consider productivity of different computers.