@buresmi7 2017-03-17
I think that main advantage is:
1) if there are not any `$excludeElements`, you do not need wait 1 second
2) 1 second is too much to wait that are all hiding block redrawed in typical case - for few block to hide - for example I use maximal 5. 

`$this->webDriverModule->wait(1);` says: Ok, for sure I will wait one second, but I don't know when redrawing is done. For few hidding block is it fine, for more, I have no idea how long it takes. In addition I don't know how fast is hardware where I running.

`$this->webDriverModule->waitForElementNotVisible($element);` says: wait for element is successfully hidden. There is no unsecurity.

So for few blocks it will be much faster. For lot of block it will be safer.