@RayRom 2017-03-17
LOL! This option is worse than waiting of 1 second after a cycle. It is possible more simply, to expect after a cycle disappearance of the last element from an array.