@buresmi7 2017-03-17
You're right, I will fix it. I did not realize, that method `executeScript` in webdriver is blocking operation