@RayRom 2017-03-17
I offer such variation:
`private function hideElementsForScreenshot(array $excludeElements)
    {
        foreach ($excludeElements as $element) {
            $this->hideElement($element);
        }
        $this->webDriverModule->waitForElementNotVisible(array_pop($element));
    }`